refactor(fe): refactor import problem modals between assignment and contest
