+++ apps/frontend/app/admin/_components/ImportProblemDescription.tsx
@@ -6,15 +6,20 @@ export function ImportProblemDescription() {
   const { table } = useDataTable()
 
   return (
-    <p className="text-primary pb-2 text-sm font-normal">
-      {table.getFilteredSelectedRowModel().rows.length} problem(s) selected:{' '}
-      {table
-        .getFilteredSelectedRowModel()
-        .rows.map((row) => {
-          const original = row.original as { title: string }
-          return original.title
-        })
-        .join(', ')}
+    <p className="pb-2 text-sm font-normal">
+      <span className="text-primary">
+        {table.getFilteredSelectedRowModel().rows.length} problem(s)
+        selected:{' '}
+      </span>
+      <span className="text-[#8A8A8A]">
+        {table
+          .getFilteredSelectedRowModel()
+          .rows.map((row) => {
+            const original = row.original as { title: string }
+            return original.title
+          })
+          .join(', ')}
+      </span>
     </p>
   )
 }
