+++ apps/frontend/app/admin/contest/_components/ImportProblemTable.tsx
@@ -2,6 +2,7 @@ import { GET_PROBLEMS } from '@/graphql/problem/queries'
 import { useSuspenseQuery } from '@apollo/client'
 import { Language, Level } from '@generated/graphql'
 import { toast } from 'sonner'
+import { ImportProblemButton } from '../../_components/ImportProblemButton'
 import { ImportProblemDescription } from '../../_components/ImportProblemDescription'
 import {
   DataTable,
@@ -12,7 +13,6 @@ import {
   DataTableSearchBar
 } from '../../_components/table'
 import type { ContestProblem } from '../_libs/schemas'
-import { ImportProblemButton } from './ImportProblemButton'
 import {
   columns,
   DEFAULT_PAGE_SIZE,
@@ -80,11 +80,15 @@ export function ImportProblemTable({
       defaultSortState={[{ id: 'select', desc: true }]}
     >
       <ImportProblemDescription />
-      <div className="flex gap-[6px] pb-1">
-        <DataTableSearchBar columndId="title" className="lg:w-[308px]" />
-        <DataTableLangFilter />
+      <div className="flex justify-between">
+        <div className="flex gap-[8px]">
+          <DataTableSearchBar columndId="title" size="sm" />
+          <DataTableLangFilter />
+        </div>
+        <ImportProblemButton onSelectedExport={onSelectedExport} />
       </div>
       <DataTable
+        size="sm"
         isModalDataTable={true}
         headerStyle={{
           select: 'rounded-l-full',
@@ -103,10 +107,8 @@ export function ImportProblemTable({
           }
         }}
       />
-      <div className="h-5" />
+      <div className="h-[12px]" />
       <DataTablePagination showRowsPerPage={false} />
-      <div className="h-7" />
-      <ImportProblemButton onSelectedExport={onSelectedExport} />
     </DataTableRoot>
   )
 }
